requirements.txt
amqp
asgiref
async-timeout
billiard
celery
sqlparse
asgiref
django
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
coreapi
coreschema
coverage
cron-descriptor
cryptography
Django
django-celery-beat
django-cors-headers
django-filter
django-stubs
django-stubs-ext
django-timezone-field
djangorestframework
djangorestframework-simplejwt
dnspython
drf-yasg
eventlet
greenlet
idna
inflection
itypes
Jinja2
kombu
MarkupSafe
mypy
mypy-extensions
packaging
Pillow
prompt-toolkit
psycopg2-binary
pycparser
PyJWT
pyTelegramBotAPI
python-crontab
python-dateutil
python-dotenv
pytz
PyYAML
redis
requests
six
sqlparse
types-pytz
types-PyYAML
typing_extensions
tzdata
uritemplate
urllib3
vine
wcwidth
